requirements.txt
# Python dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0  # Async SQLite driver
psycopg2-binary>=2.9.9  # PostgreSQL (for production)

# Machine Learning (already in use)
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
